requirements.txt
websocket-client==1.5.1
scikit-learn==1.2.2
pandas==1.5.3